Add core Android Studio files. Reorganize code. Optimize imports
