+++ app/src/main/java/com/example/laurenfalzarano/shots/Activities/AddCounterActivity.java
@@ -1,11 +1,9 @@
-package com.example.laurenfalzarano.shots;
+package com.example.laurenfalzarano.shots.Activities;
 
 import android.content.Intent;
-import android.content.SharedPreferences;
-import android.preference.PreferenceManager;
+import android.os.Bundle;
 import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
 import android.util.Log;
 import android.view.Gravity;
 import android.view.Menu;
@@ -13,13 +11,10 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.Toast;
 
+import com.example.laurenfalzarano.shots.Persistence.PreferencesManager;
+import com.example.laurenfalzarano.shots.R;
 import com.joanzapata.iconify.widget.IconTextView;
-
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.Set;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
