+++ app/src/main/java/com/example/laurenfalzarano/shots/Activities/ShotsActivity.java
@@ -1,11 +1,10 @@
-package com.example.laurenfalzarano.shots;
+package com.example.laurenfalzarano.shots.Activities;
 
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
-import android.view.View;
+
+import com.example.laurenfalzarano.shots.R;
 
 public class ShotsActivity extends AppCompatActivity {
