--- app/src/main/java/com/example/laurenfalzarano/shots/Adapters/ShotsAdapter.java
@@ -1,8 +1,0 @@
-package com.example.laurenfalzarano.shots.Adapters;
-
-/**
- * Created by laurenfalzarano on 11/7/15.
- */
-public class ShotsAdapter {
-
-}
