+++ app/src/main/java/com/example/laurenfalzarano/shots/Adapters/ShotsListAdapter.java
@@ -1,19 +1,16 @@
-package com.example.laurenfalzarano.shots;
+package com.example.laurenfalzarano.shots.Adapters;
 
 import android.content.Context;
-import android.content.SharedPreferences;
-import android.preference.PreferenceManager;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.TextView;
 
-import java.util.ArrayList;
-import java.util.Iterator;
+import com.example.laurenfalzarano.shots.Persistence.PreferencesManager;
+import com.example.laurenfalzarano.shots.R;
+
 import java.util.List;
-import java.util.Set;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
