+++ app/src/main/res/layout/activity_add_counter.xml
@@ -6,7 +6,7 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:descendantFocusability="beforeDescendants"
     android:focusableInTouchMode="true"
-    tools:context="com.example.laurenfalzarano.shots.AddCounterActivity">
+    tools:context="com.example.laurenfalzarano.shots.Activities.AddCounterActivity">
 
     <EditText
         android:layout_width="match_parent"
