+++ app/src/main/res/layout/activity_shots.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.example.laurenfalzarano.shots.ShotsActivity">
+    tools:context="com.example.laurenfalzarano.shots.Activities.ShotsActivity">
 
     <android.support.design.widget.AppBarLayout android:layout_height="wrap_content"
         android:layout_width="match_parent" android:theme="@style/AppTheme.AppBarOverlay">
