+++ app/src/main/res/layout/content_shots.xml
@@ -10,7 +10,7 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:showIn="@layout/activity_shots"
-    tools:context="com.example.laurenfalzarano.shots.ShotsActivity">
+    tools:context="com.example.laurenfalzarano.shots.Activities.ShotsActivity">
 
     <EditText
         android:layout_width="wrap_content"
